Start making test FSM
The idea here is that the toplevel of the test runner doesn't
know much about how the threads synchronise with each other;
instead, it consumes a recursive loop of test stages that usually
flow from 'runnable' to 'waiting'.

Occasionally, a thread will get the 'observable' stage, at which
point it should grab the current observation map and record the
present state into it; I've implemented this with a mutex for now,
but perhaps a proper concurrent map or a channel based approach
would be better.

Hopefully, this scheme makes sure that there can only be one thread
able to read the test environment, but keeps things open for future
exploration.