Implement checking protocol
Currently, we delegate checks to the C test stub, but only when we
notice that a new state has been entered.  This currently also
uses the existing environment, and so has to be done inside the
observation barrier.

I'll probably mess around with this setup if I realise it isn't
going to work properly.