Make bundle contain cloneable FSA set
This is a lot of indirection, but basically makes it so that the
full set of 'ready' FSA handles can be cloned at will but the
handles themselves can't once they're divided out.

In typing this, I'm realising this is kinda unsafe.  There should
not be a way to run two ReadySets in parallel.  I'll have a think
as to how to fix this.