Implement test killing
The observer thread now has responsiblity over telling all of the
other threads that the test is dead, and does so by twiddling a
boolean that every thread reads during its run cycle.

This will allow the observer thread to request that the test be
abandoned for reasons that can't otherwise be synchronised between
threads, such as signals or timeouts.