C: allow `check` to be missing
If the `check` symbol is missing in the test, we just mark the test as
having unknown outcomes for each state, like Herd7.

Currently this turns into an optionality check every time we admit a
new state.  This is perhaps a bit slow, so it'd be nice to improve this.

This currently just affects the C ABI side of things; there's no way for
the user to disable checking yet.  I suspect the answer to both of these
problems will be implementing a 'constant checker' and changing the
entry API to use a Box<dyn Checker>, but I couldn't work out how to do
that this morning.