Move shared state into FSA
This eliminates the Mutex being used to guard access into the shared
state, using the FSA to guarantee mutual exclusion instead.

This pattern of Arc<UnsafeCell> might be what we need for `Env` too,
but the rules on UnsafeCell suggest that what we're doing at the
moment would be undefined behaviour.  This'll need some forethought.

Fixes #12.