Simplify rust-level manifest API
Instead of manifests containing two separate VarMaps (per atomicity)
and having the indexing be implicit in the ordering of the map, we
now consider the atomicity to be part of the index (a 'slot') and
store the slot in the VarRecord.

This has made the construction of environments a bit less efficient;
we need to check every slot to get the maximum slot number for each
atomicity, and add one.  We should only be doing this once though.

This also paves the way for using HashMaps instead of BTreeMaps in
manifests.