Move env reference counting to Rust code
Env is now behind an Arc<UnsafeCell<Env>>.  This has a few semantic
consequences: we now consider environments being passed to the test
code as const (as any changes to them are outside our codebase and
effectively IO), and explicitly mention the invariants needed to
establish some modicum of Rust ownership model safety (ie tests running
are immutable borrows and cannot intersect with observing the env which
is a mutable borrow).

Eventually, we'll merge the Env into the test state, and that'll clean
the FSA up even more.

The C environment still has code for reference counting, for now.
This will be replaced with deep copying soon, when we consider how to
make environment files.