Overhaul configuration setup
What used to be the args struct is now a config struct, with a few
fairly hefty changes to prepare for eventually being able to parse
the config from a disk file by default.

This also implements speedchecking by virtue of broadening the config
support for checking to allow exiting on particular check outcomes.

Fixes #7.